Fixes for lobby and gamelib and uno
The lobby had a few issues around loading in the commands from launched games when the command should have been for the game.

The gamelib had a bug around not picking up old messages.

Lots of bug fixes and feature improvements to uno.  It still has some big bugs and a few missing features, but is kind of playable.